Fix reliance on using `str` for extracting an element's CDATA.
Due to an unfortunate design choice, the typical interface to extract
the CDATA inside a `domish.Element` is using `str` on the element and is
implemented by `__str__` returning the first string-like object in its
list of children, regardless of being bytes or unicode strings.
See #3579 for details.

For now, use `unicode` instead, even though the actual implementation is
in `__str__`. This bug became more obvious because of the change to
`io.StringIO` for the Python 3 port.